Bug fixes for ivars support.
Convert FFT code over to using ivars in an attempt to remove the A()
argument concatenation metaprogramming technique. There are
difficulties doing this in full because not all of the uses of A() are
concatenating arguments that were passed in from the caller. This also
exposes a couple of tricky issues with using IVars to construct
wrappers due to scheduling issues.


Former-commit-id: a9f2cae208f533605c9c10a656b4a21b60d802e8