In EliminateInterleave, handle concat_vectors of two interleaves of double vectors.
In this patch, we look for this pattern

dv2 = slice_vector(concat_vector(interleave(dv0), interleave(dv1), ..);
and we convert it into
dv2 = interleave(concat_vector(slice_vector(dv0)), slice_vector(dv1))
dv0, dv1 and dv2 are double vectors.
This helps us cancel the interleave of a double vector with a higher
(in the expression tree) deinterleave. This way we avoid having to
concatenate two double vectors and then have to slice a quad vector
while leaving ourselves with a chance to cancel the interleave with
a single deinterleave.


Former-commit-id: b9869a834c1373d445fbea7a94ae828abd1c7518